+++ test/test_agents.js
@@ -1985,5 +1985,616 @@ describe('Agents', function() {
 
 
     }); // GET/agents/stats/distinct
+    
+    
+	describe('GET/agents/:agent/config/:component/:configuration', function () {
+		
+		// agent	
+		it('Request-Agent-Client', function(done) {
+            request(common.url)
+            .get("/agents/001/config/agent/client")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+				res.body.data.should.have.properties('client');
+				res.body.data.client.should.have.properties(['crypto_method', 'remote_conf', 'auto_restart',
+				'server', 'config-profile', 'time-reconnect', 'notify_time']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Agent-Buffer', function(done) {
+            request(common.url)
+            .get("/agents/001/config/agent/buffer")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('buffer');
+				res.body.data.buffer.should.have.properties(['disabled',
+				'queue_size', 'events_per_second']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Agent-Labels', function(done) {
+            request(common.url)
+            .get("/agents/001/config/agent/labels")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                // res.body.data.shoul.have.properties('labels'); // empty list	
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+
+		it('Request-Agent-Internal', function(done) {
+            request(common.url)
+            .get("/agents/001/config/agent/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('internal');
+				res.body.data.internal.should.have.properties(['monitord', 'remoted',
+				'agent']);
+				res.body.data.internal.monitord.should.have.properties(['daily_rotations',
+				'day_wait', 'keep_log_days', 'compress', 'size_rotate', 'rotate_log']);
+				res.body.data.internal.remoted.should.have.properties('request_rto_msec',
+				'recv_counter_flush', 'request_pool', 'comp_average_printout',
+				'verify_msg_id', 'max_attempts');
+				res.body.data.internal.agent.should.have.properties('normal_level',
+				'min_eps', 'recv_timeout', 'state_interval', 'warn_level', 'debug', 'tolerance');
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+		
+		// agentless
+		it('Request-Agentless-Agentless', function(done) {
+            request(common.url)
+            .get("/agents/000/config/agentless/agentless")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('agentless'); // returns an array
+                res.body.data.agentless[0].should.have.properties(['state', 'host',
+                'frequency', 'arguments', 'type', 'port']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // analysis
+		it('Request-Analysis-Global', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/global")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['global']);
+                res.body.data.global.should.have.properties(['email_notification', 'max_output_size',
+                'alerts_log', 'zeromq_output', 'host_information', 'jsonout_output', 'rotate_interval',
+                'rootkit_detection', 'integrity_checking', 'memory_size', 'logall', 'prelude_output',
+                'stats', 'white_list', 'logall_json']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Analysis-Active-response', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/active_response")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                // res.body.data.should.have.properties(['active_response']); // empty list
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Analysis-Alerts', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/alerts")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['alerts']);
+                res.body.data.alerts.should.have.properties(['email_alert_level', 'log_alert_level']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Analysis-Command', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/command")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('command');
+                res.body.data.command[0].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[1].should.have.properties(['executable', 'timeout_allowed',
+                'name']);
+                res.body.data.command[2].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[3].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[4].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[5].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[6].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[7].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);
+                res.body.data.command[8].should.have.properties(['executable', 'timeout_allowed',
+                'name', 'expect']);                
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Analysis-Rules', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/rules")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['rules']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Analysis-Decoders', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/decoders")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('decoders');
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Analysis-Internal', function(done) {
+            request(common.url)
+            .get("/agents/000/config/analysis/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('internal');
+                res.body.data.internal.should.have.properties(['analysisd']);
+                res.body.data.internal.analysisd.should.have.properties(['label_cache_maxage',
+                'stats_percent_diff', 'show_hidden_labels', 'decoder_order_size',
+                'min_rotate_interval', 'stats_mindiff', 'log_fw', 'rlimit_nofile', 'fts_list_size',
+                'debug', 'fts_min_size_for_str', 'default_timeframe', 'stats_maxdiff']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // auth
+		it('Request-Auth-Auth', function(done) {
+            request(common.url)
+            .get("/agents/000/config/auth/auth")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('auth');
+                res.body.data.auth.should.have.properties(['purge', 'ssl_auto_negotiate', 'ciphers',
+                'force_insert', 'ssl_verify_host', 'limit_maxagents', 'force_time',
+                'ssl_manager_key', 'disabled', 'ssl_manager_cert', 'use_source_ip',
+                'use_password', 'port']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // com
+		it('Request-Com-Active-response', function(done) {
+            request(common.url)
+            .get("/agents/001/config/com/active-response")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['active-response']);
+                res.body.data['active-response'].should.have.properties(['disabled', 'ca_store']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Com-Internal', function(done) {
+            request(common.url)
+            .get("/agents/001/config/com/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['internal']);
+                res.body.data.internal.should.have.properties(['execd']);
+                res.body.data.internal.execd.should.have.properties(['request_timeout', 'max_restart_lock']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // csyslog
+		it('Request-Csyslog-Csyslog', function(done) {
+            request(common.url)
+            .get("/agents/000/config/csyslog/csyslog")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['syslog_output']);
+                res.body.data['syslog_output'][0].should.have.properties(['format',
+                'level', 'use_fqdn', 'port', 'server']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // integrator
+		it('Request-Integrator-Integration', function(done) {
+            request(common.url)
+            .get("/agents/000/config/integrator/integration")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('integration');
+                res.body.data.integration[0].should.have.properties(['alert_format', 'hook_url',
+                'group', 'name', 'level']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+
+        // logcollector  // fails without any motive
+		it('Request-Logcollector-Localfile', function(done) {
+            request(common.url)
+            .get("/agents/001/config/logcollector/localfile")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('localfile');
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Logcollector-Socket', function(done) {
+            request(common.url)
+            .get("/agents/001/config/logcollector/socket")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                // res.body.should.have.properties(['error', 'data']); // data property is empty
+                // res.body.data.should.have.properties(['socket']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Logcollector-Internal', function(done) {
+            request(common.url)
+            .get("/agents/001/config/logcollector/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('internal');
+                res.body.data.internal.should.have.properties('logcollector');
+				res.body.data.internal.logcollector.should.have.properties(['open_attempts', 'input_threads',
+                'vcheck_files', 'max_files', 'sock_fail_time', 'queue_size', 'max_lines', 'remote_commands',
+                'loop_timeout', 'debug', 'open_attempts']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // mail
+		it('Request-Mail-Global', function(done) {
+            request(common.url)
+            .get("/agents/000/config/mail/global")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('global');
+                res.body.data.global.should.have.properties(['email_maxperhour', 'email_to',
+                'email_from', 'smtp_server']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Mail-Alerts', function(done) {
+            request(common.url)
+            .get("/agents/000/config/mail/alerts")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                // res.body.should.have.properties(['error', 'data']); // data property is empty
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Mail-Internal', function(done) {
+            request(common.url)
+            .get("/agents/000/config/mail/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('internal');
+                res.body.data.internal.should.have.properties('mail');
+                res.body.data.internal.mail.should.have.properties(['strict_checking',
+                'grouping']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // monitor
+		it('Request-Monitor-Internal', function(done) {
+            request(common.url)
+            .get("/agents/000/config/monitor/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('monitord');
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // request
+		it('Request-Request-Remote', function(done) {
+            request(common.url)
+            .get("/agents/000/config/request/remote")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('remote');
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Request-Internal', function(done) {
+            request(common.url)
+            .get("/agents/000/config/request/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties('internal');
+                res.body.data.internal.should.have.properties('remoted');
+                res.body.data.internal.remoted.should.have.properties(['request_timeout', 'pass_empty_keyfile',
+                'recv_timeout', 'request_rto_sec', 'request_rto_msec', 'response_timeout', 'sender_pool', 'recv_counter_flush',
+                'request_pool', 'comp_average_printout', 'shared_reload', 'merge_shared', 'rlimit_nofile',
+                'verify_msg_id', 'max_attempts']);
+
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // syscheck
+		it('Request-Syscheck-Syscheck', function(done) {
+            request(common.url)
+            .get("/agents/001/config/syscheck/syscheck")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['syscheck']);
+                res.body.data.syscheck.should.have.properties(['ignore', 'skip_nfs', 'directories',
+                'scan_on_start', 'disabled', 'frequency', 'restart_audit', 'nodiff']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Syscheck-Rootcheck', function(done) {
+            request(common.url)
+            .get("/agents/001/config/syscheck/rootcheck")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['rootcheck']);
+                res.body.data.rootcheck.should.have.properties(['check_unixaudit', 'check_sys', 'rootkit_trojans',
+                'skip_nfs', 'check_if', 'check_pids', 'check_dev', 'check_ports', 'disabled', 'rootkit_files',
+                // 'frequency', 'scanall', 'check_trojans', 'base_directory', 'check_files', 'system_audit']); // base directory value is empty, this cause an error
+                'frequency', 'scanall', 'check_trojans', 'check_files', 'system_audit']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        it('Request-Syscheck-Internal', function(done) {
+            request(common.url)
+            .get("/agents/001/config/syscheck/internal")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['internal']);
+                res.body.data.internal.should.have.properties(['syscheck', 'rootcheck']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+        // wmodules
+		it('Request-Wmodules-Wmodules', function(done) {
+            request(common.url)
+            .get("/agents/001/config/wmodules/wmodules")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+                res.body.data.should.have.properties(['wmodules']);
+                //res.body.data.wmodules.should.have.properties(['open-scap', 'cis-cat',
+                //'osquery', 'syscollector', 'database', 'wazuh_download']);
+                res.body.data.wmodules[0].should.have.properties(['open-scap']);
+                res.body.data.wmodules[1].should.have.properties(['cis-cat']);
+                res.body.data.wmodules[2].should.have.properties(['osquery']);
+                res.body.data.wmodules[3].should.have.properties(['syscollector']);
+                res.body.data.wmodules[4].should.have.properties(['database']);
+                res.body.data.wmodules[5].should.have.properties(['wazuh_download']);
+                
+                res.body.error.should.equal(0);
+                done();
+            });
+        });
+        
+
+    }); // GET/agents/:agent/config/:component/:configuration
 
 });  // Agents